desenvolvimento do produto

+++ src/main/java/com/gerenciador_estoque/gerenciador_estoque/service/ServiceCategory.java
@@ -17,18 +17,14 @@ import jakarta.validation.Valid;
 
 @Service
 @Validated
-public class ServiceCategory implements ServiceInterface<Category, Long> {
+public class ServiceCategory implements ServiceInterface<Category, Long>{
     @Autowired
     private RepositoryCategory repositoryCategory;
 
-    @Override
+    
     public Boolean save(@Valid Category object) throws Exception {
         if (object == null) {
             throw new IllegalArgumentException("object is null");
-        }
-
-        if (repositoryCategory.findById(object.getId()) != null) {
-            throw new DataIntegrityViolationException("Category with the same ID already exists");
         }
 
         repositoryCategory.save(object);
@@ -36,12 +32,10 @@ public class ServiceCategory implements ServiceInterface<Category, Long> {
         return true;
     }
 
-    @Override
-    public Boolean update(Category object) throws Exception {
+    public Boolean update(@Valid Category object) throws Exception {
         return update(object,object.getId());
     }
 
-    @Override
     public Boolean update(@Valid Category object, Long id) throws Exception {
         if (findById(id).isEmpty()) {
             throw new EntityNotFoundException("Product with ID" + id +" not found");
@@ -55,21 +49,21 @@ public class ServiceCategory implements ServiceInterface<Category, Long> {
         repositoryCategory.save(object);
 
         return true;
-       
     }
 
-    @Override
     public Boolean delete(Long id) {
+        if (id < 0) {
+            return false;
+        }
         if (findById(id).isEmpty()){
             return false;
         }
-
+        
         repositoryCategory.deleteById(id);
 
         return true;
     }
 
-    @Override
     public Optional<Category> findById(Long id)  {
         if (id < 0 ) {
             return Optional.empty();
@@ -78,7 +72,6 @@ public class ServiceCategory implements ServiceInterface<Category, Long> {
         return repositoryCategory.findById(id);
     }
 
-    @Override
     public List<Category> findAll() {
         return repositoryCategory.findAll();
     }
